requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
scikit-learn==1.5.2
pandas==2.2.2
numpy==2.1.1
streamlit==1.38.0
plotly==5.24.1
APScheduler==3.10.4
requests
python-dotenv
gunicorn